[refactor/#668] ChatQueryServiceImpl에서 party/direct 목록 조회 책임 분리 (#681)
* refactor: 모임 채팅방 조회 로직을 별도 서비스로 분리

* refactor: 개인 채팅방 조회 로직을 별도 서비스로 분리

* refactor: 불필요한 import 제거 및 위임 로직 단위 테스트 추가

* refactor: command/query 패키지로 정리

* refactor: ImageUrlResolver를 사용하도록 변경

* refactor: 알 수 없는 사용자 프로필 메서드명 명확화

* fix: 모임 채팅방 조회에서 불필요한 중복 조회 로직을 제거
