+++ src/main/java/umc/cockple/demo/domain/chat/service/query/ChatQueryService.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.domain.chat.service;
+package umc.cockple.demo.domain.chat.service.query;
 
 import umc.cockple.demo.domain.chat.dto.*;
